+++ bindings/pydeck/pydeck/__init__.py
@@ -1,7 +1,5 @@
 from .bindings import map_styles, Deck, Effect, Extension, Layer, View, ViewState, Widget  # noqa
 
-from .nbextension import _jupyter_nbextension_paths  # noqa
-
 from ._version import __version__  # noqa
 
 from .settings import settings  # noqa